Rename /server/typings to /server/types/
