+++ server/src/features/game/gameUtils.ts
@@ -1,7 +1,7 @@
 import _ from "lodash";
 import dayjs, { Dayjs } from "dayjs";
 import { findGames, removeGames } from "server/features/game/gameRepository";
-import { GameDoc } from "server/typings/gameTypes";
+import { GameDoc } from "server/types/gameTypes";
 import { logger } from "server/utils/logger";
 import {
   Game,
