+++ server/src/features/player-assignment/padg/utils/assignPadg.ts
@@ -6,11 +6,11 @@ import {
   Group,
   Event,
   PadgRandomAssignResults,
-} from "server/typings/padgRandomAssignTypes";
+} from "server/types/padgRandomAssignTypes";
 import { logger } from "server/utils/logger";
 import { config } from "shared/config";
 import { calculateHappiness } from "server/features/player-assignment/padg/utils/calculateHappiness";
-import { Input } from "server/typings/resultTypes";
+import { Input } from "server/types/resultTypes";
 
 export const assignPadg = (
   groups: Group[],
