+++ server/src/features/player-assignment/utils/formatResults.ts
@@ -1,5 +1,5 @@
 import _ from "lodash";
-import { PadgRandomAssignResults } from "server/typings/padgRandomAssignTypes";
+import { PadgRandomAssignResults } from "server/types/padgRandomAssignTypes";
 import { logger } from "server/utils/logger";
 import { AssignmentError } from "shared/typings/api/errors";
 import { AssignmentResult } from "shared/typings/models/result";
