+++ server/src/features/player-assignment/utils/getGroups.ts
@@ -1,6 +1,6 @@
 import dayjs from "dayjs";
 import _ from "lodash";
-import { Group } from "server/typings/padgRandomAssignTypes";
+import { Group } from "server/types/padgRandomAssignTypes";
 import { logger } from "server/utils/logger";
 import { AssignmentError } from "shared/typings/api/errors";
 import { User } from "shared/typings/models/user";
