+++ server/src/features/statistics/fixer-helpers/dataAnonymizer.ts
@@ -2,7 +2,7 @@ import fs from "fs";
 import { faker } from "@faker-js/faker";
 import { logger } from "server/utils/logger";
 import { User } from "shared/typings/models/user";
-import { ResultsCollectionEntry } from "server/typings/resultTypes";
+import { ResultsCollectionEntry } from "server/types/resultTypes";
 import { writeJson } from "server/features/statistics/statsUtil";
 import { config } from "shared/config";
 import { Signup } from "server/features/signup/signupTypes";
