+++ server/src/features/statistics/fixer-helpers/gameIdFix.ts
@@ -2,13 +2,13 @@ import fs from "fs";
 import _ from "lodash";
 import { logger } from "server/utils/logger";
 import { User } from "shared/typings/models/user";
-import { GameDoc } from "server/typings/gameTypes";
-import { ResultsCollectionEntry } from "server/typings/resultTypes";
+import { GameDoc } from "server/types/gameTypes";
+import { ResultsCollectionEntry } from "server/types/resultTypes";
 import { writeJson } from "server/features/statistics/statsUtil";
 import { config } from "shared/config";
 import { Game } from "shared/typings/models/game";
 import { SignupDoc } from "server/features/signup/signupTypes";
-import { SettingsDoc } from "server/typings/settingsTypes";
+import { SettingsDoc } from "server/types/settingsTypes";
 
 export const gameIdFix = async (year: number, event: string): Promise<void> => {
   const users: User[] = JSON.parse(
