+++ server/src/features/statistics/statistics-helpers/gameDataHelpers.ts
@@ -7,7 +7,7 @@ import { getMaximumNumberOfPlayersByTime } from "./resultDataHelpers";
 import {
   StringNumberObject,
   PriorityObject,
-} from "server/typings/commonTypes";
+} from "server/types/commonTypes";
 import { toPercent } from "server/features/statistics/statsUtil";
 import { TIMEZONE } from "shared/utils/initializeDayjs";
