+++ server/src/features/statistics/statistics-helpers/resultDataHelpers.ts
@@ -1,7 +1,7 @@
 import { logger } from "server/utils/logger";
 import { Game } from "shared/typings/models/game";
-import { ResultsCollectionEntry } from "server/typings/resultTypes";
-import { StringNumberObject } from "server/typings/commonTypes";
+import { ResultsCollectionEntry } from "server/types/resultTypes";
+import { StringNumberObject } from "server/types/commonTypes";
 import { toPercent } from "server/features/statistics/statsUtil";
 
 export const getSignupsByTime = (
