+++ server/src/features/user/userUtils.ts
@@ -1,7 +1,7 @@
 import { Dayjs } from "dayjs";
 import { logger } from "server/utils/logger";
 import { saveSerials } from "server/features/serial/serialRepository";
-import { SerialDoc } from "server/typings/serialTypes";
+import { SerialDoc } from "server/types/serialTypes";
 import { Result } from "shared/utils/result";
 import { MongoDbError } from "shared/typings/api/errors";
