+++ server/src/types/commonTypes.ts
