+++ server/src/types/declarations/eventassigner-js.d.ts
@@ -2,7 +2,7 @@ declare module "eventassigner-js" {
   import {
     Input,
     PadgAssignResults,
-  } from "server/typings/padgRandomAssignTypes";
+  } from "server/types/padgRandomAssignTypes";
   const defaultImport: { eventAssignment: (input: Input) => PadgAssignResults };
   export = defaultImport;
 }
