+++ server/src/types/declarations/generate-serial-number.d.ts
