+++ server/src/types/resultTypes.ts
