+++ server/src/types/serialTypes.ts
