+++ server/src/types/settingsTypes.ts
